% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize.R
\name{compile_custom_fitness}
\alias{compile_custom_fitness}
\title{custom_fitness_distribution}
\usage{
compile_custom_fitness(cppfile)
}
\arguments{
\item{cppfile}{a character string giving the path name of a cpp file}
}
\description{
Allows the user to create a custom fitness distribution in C++ for use in
SIApop. Creates the appropriate header file and compiles the source file then
links into a shared object. The created shared object is used as a plugin.
}
\examples{
\dontrun{
compile_custom_fitness(cppfile = "./plugin.cpp")
}
}
