% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize.R
\name{compile_custom_newclone}
\alias{compile_custom_newclone}
\title{custom_custom_newclone}
\usage{
compile_custom_newclone(cppfile)
}
\arguments{
\item{cppfile}{a character string giving the path name of a cpp file}
}
\description{
Allows the user to create a custom plugin in C++ for use in
SIApop. The plugin would allow new inheritance scenarios for a new clone
enterring the population.
}
\examples{
\dontrun{
compile_custom_newclone(cppfile = "./plugin.cpp")
}
}

