% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_fishplot}
\alias{convert_fishplot}
\title{Converts data to fish object in package fishplot for plotting}
\usage{
convert_fishplot(time_data, threshold = 1e-04, timepoints = NULL)
}
\arguments{
\item{time_data}{dataframe of time course data}

\item{threshold}{minimum allele frequency to count a clone at}

\item{timepoints}{timepoints to use}
}
\value{
convert_fishplot - returns a fish object for use with fishPlot()
}
\description{
Converts data to fish object in package fishplot for plotting
}

