% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_igraph}
\alias{convert_igraph}
\title{Converts data to igraph in tree layout to view the genetree at a specific time}
\usage{
convert_igraph(clonedata, threshold = 0.01, size = NULL, color = NULL)
}
\arguments{
\item{clonedata}{dataframe of a single timepoint}

\item{threshold}{minimum allele frequency to count a clone at}
}
\value{
convert_igraph - returns an igraph object
}
\description{
Converts data to igraph in tree layout to view the genetree at a specific time
}

