% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize.R
\name{create_fitness_template}
\alias{create_fitness_template}
\title{create_fitness_template}
\usage{
create_fitness_template(cppfile = "custom_dist_plugin.cpp")
}
\arguments{
\item{cppfile}{file name of .cpp file to create}
}
\description{
Creates a fitness distribution ".cpp" template file at the given location
which can be modified before running `compile_custom_fitness`.
}
\examples{
\dontrun{
create_fitness_template(cppfile = "custom_dist_plugin.cpp")
}
}

