% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_Muller_df}
\alias{create_Muller_df}
\title{Wrapper to convert from list containing edgelist and pop_df to Muller_df}
\usage{
create_Muller_df(Muller_list)
}
\arguments{
\item{Muller_list}{list as output from convert_ggmuller containing edge list and pop_df}
}
\value{
create_Muller_df - returns a Muller_df tibble for Muller_plot
}
\description{
Wrapper to convert from list containing edgelist and pop_df to Muller_df
}

