% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize.R
\name{create_newclone_template}
\alias{create_newclone_template}
\title{create_newclone_template}
\usage{
create_newclone_template(cppfile = "custom_newclone_plugin.cpp")
}
\arguments{
\item{cppfile}{file name of .cpp file to create}
}
\description{
Creates a new clone plugin ".cpp" template file at the given location
which can be modified before running `compile_custom_newclone`.
}
\examples{
\dontrun{
create_newclone_template(cppfile = "custom_newclone_plugin.cpp")
}
}
