% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_sample_adj_matrix}
\alias{create_sample_adj_matrix}
\title{Creates an adjacency matrix from a sample from SIApop}
\usage{
create_sample_adj_matrix(samp)
}
\arguments{
\item{samp}{SIApop single sample}
}
\value{
create_sample_adj_matrix - returns an adjacency matrix for a sample
}
\description{
Creates an adjacency matrix from a sample from SIApop
}

