% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.create_edge.list}
\alias{.create_edge.list}
\title{.create_edge_list}
\usage{
.create_edge.list(id_list, reduce = TRUE)
}
\arguments{
\item{id_list}{list of clone ids}

\item{reduce}{if true we label a clone by its number rather than the list of
ancestors}
}
\value{
.create_edge.list - returns an edge list for the population
}
\description{
Converts data to Muller_df object in package ggmuller for plotting
}
