% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.match_parents}
\alias{.match_parents}
\title{.match_parents}
\usage{
.match_parents(ids)
}
\arguments{
\item{ids}{- vector of allele ids.}
}
\value{
.match_parents - returns a vector of parent indices where 0 indicates
 an ancestor
}
\description{
matches an allele index to the respective parent index
}
