% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.parent_age_at_split}
\alias{.parent_age_at_split}
\title{.parent_age_at_split}
\usage{
.parent_age_at_split(Identity, edgelist, timevec)
}
\arguments{
\item{Identity}{character vector of identity of a clone}

\item{edgelist}{edgelist mapping parents to clones}

\item{timevec}{vector of birth times of each clone in the same order as edgelist}
}
\value{
.parent_age_at_split - returns a vector of times for parents ages at split
}
\description{
Gets the parents age of an individual from an edgelist and the vector of birth times
}
