% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.replace}
\alias{.replace}
\title{.replace}
\usage{
.replace(vec, pattern, replacement)
}
\arguments{
\item{vec}{- character vector}

\item{pattern}{- character element to search for}

\item{replacement}{- character to replace pattern with}
}
\value{
.replace - returns a vector with the replaced elements
}
\description{
replaces an element of a vector with a particular pattern with the
replacement given
}
