% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.split_replace_collapse}
\alias{.split_replace_collapse}
\title{.split_replace_collapse}
\usage{
.split_replace_collapse(vec, pattern, replacement)
}
\arguments{
\item{vec}{- character vector of ids}

\item{pattern}{- character element to search for}

\item{replacement}{- character to replace pattern with}
}
\value{
.split_replace_collapse - returns a vector with the replaced ids
}
\description{
for a list of clone ids each separated by ">" a specific allele can be
searched for and replaced by splitting the elements of a vector,
replacing individual elements, and collapsing the list back to a vector
}
