% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_clonedata}
\alias{import_clonedata}
\title{Import clone data as list of data frames by simulation number}
\usage{
import_clonedata(filedir)
}
\arguments{
\item{filedir}{folder directory containing clonedata.txt}
}
\value{
clone_data - A list of data frames for each simulation
}
\description{
Import clone data as list of data frames by simulation number
}

