% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_sampledata}
\alias{import_sampledata}
\title{Import single cell samples}
\usage{
import_sampledata(filename = "./sampledata.txt")
}
\arguments{
\item{filedir}{folder directory containing clonedata.txt}
}
\value{
sim_data - A list of variables for all simulations
}
\description{
Import single cell samples
}

