% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_simdata}
\alias{import_simdata}
\title{Import summary of simulations and convert to a list}
\usage{
import_simdata(filedir)
}
\arguments{
\item{filedir}{folder directory containing clonedata.txt}
}
\value{
sim_data - A list of variables for all simulations
}
\description{
Import summary of simulations and convert to a list
}

