% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_timedata}
\alias{import_timedata}
\title{Import time course data as list of data frames by simulation number}
\usage{
import_timedata(filedir)
}
\arguments{
\item{filedir}{folder directory containing clonedata.txt}
}
\value{
time_data - A list of data frames for each simulation
}
\description{
Import time course data as list of data frames by simulation number
}

