% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reload}
\alias{reload}
\title{reload}
\usage{
reload(path)
}
\arguments{
\item{path}{path to siapopr package source}
}
\description{
Recompiles and reinstalls the package for use after adjusting the model
code. WARNING: For advanced users. Adjusting the C++ code can lead to
errors in compilation.
}
