% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_alleles}
\alias{sample_alleles}
\title{collect a sample of alleles from the entire population}
\usage{
sample_alleles(clone_data, n_alleles, threshold = 0.01)
}
\arguments{
\item{clone_data}{- clone data frame}

\item{n_alleles}{- number of alleles to sample total}

\item{threshold}{- minimum frequency to allow for sampling}
}
\value{
data frame of a sample of alleles
}
\description{
collect a sample of alleles from the entire population
}

