% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_cells}
\alias{sample_cells}
\title{sample cells from clone data frame (for small number of clones)}
\usage{
sample_cells(clone_data, n_cells)
}
\arguments{
\item{clone_data}{- clone data frame}

\item{n_cells}{- number of alleles to sample total}
}
\value{
data frame containing a list of clones with respective cells sampled
}
\description{
sample cells from clone data frame (for small number of clones)
}

