% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_ggmuller}
\alias{convert_ggmuller}
\title{Converts data to Muller_df object in package ggmuller for plotting}
\usage{
convert_ggmuller(time_data, threshold = 0.001, timepoints = NULL,
  freqplot = FALSE, reduce = TRUE, ...)
}
\arguments{
\item{time_data}{dataframe of time course data}

\item{threshold}{minimum allele frequency to count a clone at}

\item{timepoints}{timepoints to use}

\item{freqplot}{if TRUE then plot as a frequency relative to each time point}

\item{reduce}{if TRUE then reduce clone names to their final ID rather than their ancestry}
}
\value{
convert_fishplot - returns a fish object for use with fishPlot()
}
\description{
Converts data to Muller_df object in package ggmuller for plotting
}

