% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_adj_matrix}
\alias{create_adj_matrix}
\title{Creates an adjacency matrix from a population run from SIApop}
\usage{
create_adj_matrix(clone_data)
}
\arguments{
\item{clone_data}{SIApop single simulation}
}
\value{
create_sample_adj_matrix - returns an adjacency matrix for a sample
}
\description{
Creates an adjacency matrix from a population run from SIApop
}

