% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_sample_adj_matrix}
\alias{create_sample_adj_matrix}
\title{create_sample_adj_matrix}
\usage{
create_sample_adj_matrix(samp)
}
\arguments{
\item{samp}{SIApop single sample data frame containing a column for the
unique id (\code{unique_id}) and number of samples for each id
(\code{number_obs}).}
}
\value{
create_sample_adj_matrix - returns an adjacency matrix for a sample
}
\description{
Creates an adjacency matrix from a sample of SIApop. The adjacency matrix
assumes rows are individuals and columns are alleles and returns a TRUE
in the i,j element of the matrix if individual sample i contains allele j.
}
\details{
The adjacency matrix create allows conversion a distance matrix for the
creation of phylogenetic trees with package \code{ape} or converted to a
\code{phyDat} object in the \code{phanghorn} package.
}
\examples{
\dontrun{
# simulate process and import
siapopConstant(seed = 17, max_pop = 1000, mutation_prob = 0.05,
               observation_frequency = 1, detection_threshold = 0.005,
               num_samples = 1, sample_size = 100)
samp_data <- import_sampledata('./sampledata.txt')$`1`
create_sample_adj_matrix(samp_data)
}
}

