% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_sampledata}
\alias{import_sampledata}
\title{Import single cell samples}
\usage{
import_sampledata(filename = "./sampledata.txt")
}
\arguments{
\item{filedir}{folder directory containing clonedata.txt}
}
\value{
sim_data - A list of variables for all simulations
}
\description{
Import single cell samples
}
\examples{
\dontrun{
siapop(seed = 17, max_pop = 1000, mutation_prob = 0.05,
               observation_frequency = 1, detection_threshold = 0.005,
               num_samples = 1, sample_size = 100)
sample_data <- import_sampledata('./sampledata.txt')
}
}
