% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_siapop}
\alias{import_siapop}
\title{Import all data from a simulation and outputs a list}
\usage{
import_siapop(filedir, clonedata_file = NULL, timedata_file = NULL,
  sampledata_file = NULL, simdata_file = NULL)
}
\arguments{
\item{filedir}{folder directory containing siapop output}
}
\value{
siapop_data - Complete output of siapop in list format
}
\description{
Import all data from a simulation and outputs a list
}

