% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_simdata}
\alias{import_simdata}
\title{Import summary of simulations and convert to a list}
\usage{
import_simdata(filename = "./sim_stats.txt")
}
\arguments{
\item{filedir}{folder directory containing clonedata.txt}
}
\value{
sim_data - A list of variables for all simulations
}
\description{
Import summary of simulations and convert to a list
}
\examples{
\dontrun{
siapop(seed = 17, max_pop = 1000, mutation_prob = 0.05,
               observation_frequency = 1, detection_threshold = 0.005,
               num_samples = 1, sample_size = 100)
time_data <- import_simdata('./sim_stats.txt')
}
}
