% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_timedata}
\alias{import_timedata}
\title{import_timedata}
\usage{
import_timedata(filename = "./timedata.txt")
}
\arguments{
\item{filedir}{folder directory containing clonedata.txt}
}
\value{
time_data - A list of data frames for each simulation
}
\description{
Import time course data as list of data frames by simulation number
}
\examples{
\dontrun{
siapopConstant(seed = 17, max_pop = 1000, mutation_prob = 0.05,
               observation_frequency = 1, detection_threshold = 0.005,
               num_samples = 1, sample_size = 100)
time_data <- import_timedata('./timedata.txt')
}
}

