% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{siapopSimple}
\alias{siapopSimple}
\title{SIApop for non-mutating processes}
\usage{
siapopSimple(input = NULL, output_dir = NULL, ancestor_file = NULL)
}
\arguments{
\item{input}{input character vector of input file}

\item{output_dir}{input character vector of output location}

\item{ancestor_file}{input character vector of ancestor file}
}
\description{
SIApop for non-mutating processes
}

